requirements.txt
opencv-python==4.5.4.58
opencv-contrib-python==4.5.4.58
Pillow==8.4.0
prettytable==2.4.0